@@ -45,13 +45,13 @@ apply plugin: 'com.gorylenko.gradle-git-properties'
 
 group = 'com.checkmarx'
 version = file('version.txt').text.trim()
-sourceCompatibility = 1.8
+sourceCompatibility = 17
 def cucumberVersion = '5.7.0'
 def junitVersion = '5.7.0'
 
 repositories {
     mavenLocal()
-	mavenCentral()
+    mavenCentral()
     maven {
         url = 'https://packages.atlassian.com/mvn/maven-external'
     }
