@@ -25,7 +25,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -563,6 +563,7 @@ public class ScanUtils {
                     .finding(f)
                     .vulnerabilityPackage(getScaPackageByFinding(scaResults.getPackages(), f))
                     .vulnerabilityLink(constructVulnerabilityUrl(scaResults.getWebReportLink(), f))
+                    .scanTags(scaResults.getScanTags())
                     .build();
 
             scaDetailsList.add(scaDetails);
