@@ -293,7 +293,10 @@ public class ADOController extends AdoControllerBase {
                     .organizationId(determineNamespace(resourceContainers))
                     .gitUrl(gitUrl)
                     .build();
-
+            if(body.getResource().getCommits()!=null)
+            {
+                request.setLatestCommitterEmail(body.getResource().getCommits().get(0).getAuthor().getEmail());
+            }
             setScmInstance(controllerRequest, request);
             request.putAdditionalMetadata(ADOService.PROJECT_SELF_URL, getTheProjectURL(body.getResourceContainers()));
             addMetadataToScanRequest(adoDetailsRequest, request);
